# openapi.yaml
openapi: 3.1.0
info:
  title: Midas Match API
  version: 1.0.0
  description: |
    Job matching and career analysis platform. Provides resume parsing, AI-powered
    job matching, deep job analysis, cover letter generation, and a token-based
    usage system with Razorpay payments.

    ## Authentication
    Most endpoints use [Clerk](https://clerk.com) session authentication via cookies.
    Admin endpoints additionally verify the user ID against server-side allow-lists.

    ## Rate Limiting
    Rate limits use a sliding-window algorithm backed by Upstash Redis.
    When a limit is exceeded the response includes a `Retry-After` header (seconds).

    ## Free Tier & Tokens
    | Resource | Free Allowance | Token Cost |
    |----------|---------------|------------|
    | Job scans | 5 per day | 1 token after limit |
    | Deep scans | 3 lifetime | 1 token after limit |
    | Midas scans | 1 per week | 2 tokens after limit |
    | Cover letters | 15 per hour | — |
    | Resume parses | 10 per hour | — |

  contact:
    name: Midas Match
  license:
    name: Proprietary
servers:
  - url: https://midasmatch.app
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/analyze-job:
    post:
      operationId: analyzeJob
      summary: Deep-scan a single job listing
      description: |
        Uses an LLM to produce a detailed fit analysis between a job and a user
        profile. First 3 deep scans are free; subsequent scans cost 1 token.
        Admin users bypass all limits.
      tags: [Jobs]
      security:
        - clerkAuth: []
      x-rate-limit: 30 requests per minute per user
      x-max-duration: 30s
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeJobRequest"
      responses:
        "200":
          description: Analysis completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  requiresAuth:
                    type: boolean
                    const: true
        "403":
          description: Token balance exhausted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaywallError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/cover-letter:
    post:
      operationId: generateCoverLetter
      summary: Generate an AI cover letter
      description: |
        Produces a concise two-paragraph cover letter (~70-90 words) tailored to
        the job and user profile. Rate-limited to 15 requests per hour.
      tags: [Jobs]
      security:
        - clerkAuth: []
      x-rate-limit: 15 requests per hour per user
      x-max-duration: 15s
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterRequest"
      responses:
        "200":
          description: Cover letter generated
          content:
            application/json:
              schema:
                type: object
                required: [letter]
                properties:
                  letter:
                    type: string
                    description: Generated cover letter text
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/credit:
    post:
      operationId: adminCreditTokens
      summary: Credit tokens to a user by email
      description: Admin-only endpoint to add tokens to any user's balance.
      tags: [Admin]
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, amount]
              properties:
                email:
                  type: string
                  format: email
                  description: Target user's email
                amount:
                  type: integer
                  minimum: 1
                  description: Number of tokens to credit
      responses:
        "200":
          description: Tokens credited
          content:
            application/json:
              schema:
                type: object
                required: [success, message, newBalance]
                properties:
                  success:
                    type: boolean
                    const: true
                  message:
                    type: string
                  newBalance:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/flags:
    get:
      operationId: getFeatureFlags
      summary: Get all feature flag states
      description: Returns the current state of all feature flags including Redis overrides.
      tags: [Admin]
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Current flag values
          content:
            application/json:
              schema:
                type: object
                required: [flags]
                properties:
                  flags:
                    $ref: "#/components/schemas/FeatureFlags"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: setFeatureFlag
      summary: Set or clear a feature flag override
      description: |
        Sets a runtime Redis override for a feature flag. Pass `null` as the value
        to clear the override and revert to the environment variable default.
      tags: [Admin]
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flag, value]
              properties:
                flag:
                  type: string
                  enum:
                    - ADVANCED_FILTERS
                    - SALARY_FILTER
                    - COMPANY_SIZE_FILTER
                    - MULTI_REGION
                value:
                  type: boolean
                  nullable: true
                  description: Set to null to clear the override
      responses:
        "200":
          description: Flag updated
          content:
            application/json:
              schema:
                type: object
                required: [updated, value, currentState]
                properties:
                  updated:
                    type: string
                  value:
                    type: boolean
                    nullable: true
                  currentState:
                    $ref: "#/components/schemas/FeatureFlags"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    clerkAuth:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie
  schemas:
...
    AnalyzeJobRequest:
      type: object
      required: [job, profile]
      properties:
        job:
          type: object
          properties:
            title:
              type: string
            company:
              type: string
            location:
              type: string
            summary:
              type: string
              description: Job description (or use "description" field)
          additionalProperties: true
        profile:
          type: object
          properties:
            headline:
              type: string
            experience_years:
              type: number
            skills:
              type: array
              items:
                type: string
          additionalProperties: true
        apiKeys:
          type: object
          properties:
            OPENROUTER_API_KEY:
              type: string
            OPENAI_API_KEY:
              type: string
    AnalyzeJobResponse:
      type: object
      required: [analysis]
      properties:
        analysis:
          type: object
          required: [fit_score, strong_signals, gaps, salary_estimate, verdict]
          properties:
            fit_score:
              type: integer
              minimum: 0
              maximum: 100
              description: Overall fit score (0-100)
            strong_signals:
              type: array
              items:
                type: string
              description: Matching strengths
            gaps:
              type: array
              items:
                type: string
              description: Areas of mismatch
            salary_estimate:
              type: string
              description: Estimated salary range
            verdict:
              type: string
              description: Summary recommendation
    CoverLetterRequest:
      type: object
      required: [job, profile]
      properties:
        job:
          type: object
          properties:
            title:
              type: string
            company:
              type: string
            summary:
              type: string
              description: Job description (truncated to 1500 chars server-side)
        profile:
          type: object
          properties:
            name:
              type: string
            headline:
              type: string
            skills:
              type: array
              items:
                type: string
              maxItems: 20
        apiKey:
          type: string
          description: Optional OpenRouter API key
...
    FeatureFlags:
      type: object
      properties:
        ADVANCED_FILTERS:
          type: boolean
        SALARY_FILTER:
          type: boolean
        COMPANY_SIZE_FILTER:
          type: boolean
        MULTI_REGION:
          type: boolean
...
    PaywallError:
      type: object
      properties:
        error:
          type: string
        paywalled:
          type: boolean
          const: true
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    RateLimitError:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: integer
          description: Seconds until the rate limit resets
        rateLimited:
          type: boolean
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"